work in progress

+++ src/incatch/Edge.java
@@ -5,19 +5,19 @@ class Edge {
     private final String id;
     private final Node source;
     private final Node destination;
-    private final int weight;
+    private final double weight;
 	
     public Edge(int ii, int jj) {
 	i = ii;
 	j = jj;	    
         //System.out.println("Sono nell'Edge");
-        id=null;
-        source=null;
-        weight=0;
-        destination=null;
+        id = null;
+        source = null;
+        weight = 0;
+        destination = null;
     }
     
-     public Edge(String id, Node source, Node destination, int weight) {
+     public Edge(String id, Node source, Node destination, double weight) {
         this.id = id;
         this.source = source;
         this.destination = destination;
@@ -46,7 +46,7 @@ class Edge {
     public Node getSource() {
         return source;
     }
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
